% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_R6.R
\name{BaseBootstrap}
\alias{BaseBootstrap}
\title{Class: Bootstrap Base}
\description{
An R6 class providing the foundational logic for temperature-based bootstrapping.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The input dataframe.}

\item{\code{y_var}}{Character. The name of the dependent variable.}

\item{\code{x_vars}}{Character vector. The names of the independent variables.}

\item{\code{boot_results}}{List. Stores the results of each bootstrap round.}

\item{\code{aggregated_data}}{Dataframe. The processed results (medians, CIs).}

\item{\code{x_range}}{Numeric vector. The min and max range for simulation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseBootstrap-new}{\code{BaseBootstrap$new()}}
\item \href{#method-BaseBootstrap-clone}{\code{BaseBootstrap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseBootstrap-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseBootstrap-new}{}}}
\subsection{Method \code{new()}}{
Create a new BaseBootstrap object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBootstrap$new(data, y_var, x_vars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe.}

\item{\code{y_var}}{Name of the response variable.}

\item{\code{x_vars}}{Vector of regressor names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseBootstrap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseBootstrap-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBootstrap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
