% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_R6.R
\name{CTRF_Model}
\alias{CTRF_Model}
\title{Class: CTRF Model}
\description{
Inherits from \link{BaseBootstrap} to handle Covariate-dependent TRF models.
}
\section{Super class}{
\code{\link[CTRF:BaseBootstrap]{CTRF::BaseBootstrap}} -> \code{CTRF_Model}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{var_groups}}{List of character vectors defining variable subsets.}

\item{\code{sim_matrices}}{List of simulation matrices for each group.}

\item{\code{orders}}{List of order vectors for each group.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CTRF_Model-new}{\code{CTRF_Model$new()}}
\item \href{#method-CTRF_Model-create_sim_matrices}{\code{CTRF_Model$create_sim_matrices()}}
\item \href{#method-CTRF_Model-run_bootstrap}{\code{CTRF_Model$run_bootstrap()}}
\item \href{#method-CTRF_Model-aggregate_results}{\code{CTRF_Model$aggregate_results()}}
\item \href{#method-CTRF_Model-plot}{\code{CTRF_Model$plot()}}
\item \href{#method-CTRF_Model-clone}{\code{CTRF_Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CTRF_Model-new"></a>}}
\if{latex}{\out{\hypertarget{method-CTRF_Model-new}{}}}
\subsection{Method \code{new()}}{
Initialize CTRF model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CTRF_Model$new(data, y_var, x_vars, var_groups, orders)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe.}

\item{\code{y_var}}{Response variable name.}

\item{\code{x_vars}}{All covariate names.}

\item{\code{var_groups}}{List of character vectors for variable grouping.}

\item{\code{orders}}{List of numeric vectors c(p, q) for each group.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CTRF_Model-create_sim_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-CTRF_Model-create_sim_matrices}{}}}
\subsection{Method \code{create_sim_matrices()}}{
Create simulation matrices for all variable groups.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CTRF_Model$create_sim_matrices(x_range, t_range)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_range}}{Numeric vector for temperature grid.}

\item{\code{t_range}}{Numeric vector for scaling.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CTRF_Model-run_bootstrap"></a>}}
\if{latex}{\out{\hypertarget{method-CTRF_Model-run_bootstrap}{}}}
\subsection{Method \code{run_bootstrap()}}{
Runs the bootstrap resampling for CTRF.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CTRF_Model$run_bootstrap(rounds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rounds}}{Integer. Number of iterations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CTRF_Model-aggregate_results"></a>}}
\if{latex}{\out{\hypertarget{method-CTRF_Model-aggregate_results}{}}}
\subsection{Method \code{aggregate_results()}}{
Aggregates results for all groups.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CTRF_Model$aggregate_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CTRF_Model-plot"></a>}}
\if{latex}{\out{\hypertarget{method-CTRF_Model-plot}{}}}
\subsection{Method \code{plot()}}{
Plots partial effects for each covariate group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CTRF_Model$plot(save_dir, x_breaks = seq(-5, 35, 5), hrsuffix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_dir}}{Directory to save images.}

\item{\code{x_breaks}}{Breaks for the x-axis.}

\item{\code{hrsuffix}}{String suffix for file names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CTRF_Model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CTRF_Model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CTRF_Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
