% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_describ.R
\name{draw_tempplot}
\alias{draw_tempplot}
\title{Draw hourly plot
This could be used as "base_plot" in TRF_Model environment, so it is not automatically saved.}
\usage{
draw_tempplot(data, hour, y_var, y_lab, color_hour, x_lim, y_lim)
}
\arguments{
\item{data}{Dataframe. Usually the main dataset, NOT bootstrap results.}

\item{hour}{Numeric. Hour.}

\item{y_var}{Character. Outcome variable name.}

\item{y_lab}{Character. See https://ggplot2.tidyverse.org/reference/labs.}

\item{color_hour}{Character. Color name specified for the selected Hour.}

\item{x_lim}{Numeric vector. See below}

\item{y_lim}{Numeric vector.
See https://ggplot2.tidyverse.org/reference/scale_continuous.html#arg-limits.
Fixed limits make it easier to compare hours.}
}
\value{
An ggplot object.
}
\description{
Draw hourly plot
This could be used as "base_plot" in TRF_Model environment, so it is not automatically saved.
}
