% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_postprocess.R
\name{fff33}
\alias{fff33}
\title{Add Fourier functional terms to an existing Dataframe
with small and fixed Degrees (3).}
\usage{
fff33(data, colname, a = NULL, b = NULL)
}
\arguments{
\item{data}{Dataframe. The main dataset to work on.}

\item{colname}{Character. Name of the column.}

\item{a}{Numeric. Minimum of the column specified.}

\item{b}{Numeric. Maximum of the column specified.}
}
\value{
An Dataframe augmented with new columns.
}
\description{
Add Fourier functional terms to an existing Dataframe
with small and fixed Degrees (3).
}
