% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_postprocess.R
\name{fff55}
\alias{fff55}
\title{Add Fourier functional terms to an existing dataframe
with small and fixed Degrees (5).
Degree p or q>=3 is rarely used in the context of temperature--elec.demand.}
\usage{
fff55(data, colname, a = NULL, b = NULL)
}
\arguments{
\item{data}{Dataframe. The main dataset to work on.}

\item{colname}{Character. Name of the column.}

\item{a}{Numeric. Minimum of the column specified.}

\item{b}{Numeric. Maximum of the column specified.}
}
\value{
An Dataframe augmented with new columns.
}
\description{
Add Fourier functional terms to an existing dataframe
with small and fixed Degrees (5).
Degree p or q>=3 is rarely used in the context of temperature--elec.demand.
}
