% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_postprocess.R
\name{getendummies}
\alias{getendummies}
\title{Add dummy vairables to an existing dataframe.}
\usage{
getendummies(data, colname)
}
\arguments{
\item{data}{Dataframe. The main dataset to work on.}

\item{colname}{Character. Name of the column by which you create dummy variables.}
}
\value{
An Dataframe augmented with new columns.
}
\description{
Add dummy vairables to an existing dataframe.
}
