% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_postprocess.R
\name{nmlize}
\alias{nmlize}
\title{Normalize an vector}
\usage{
nmlize(x, a = NULL, b = NULL)
}
\arguments{
\item{x}{Vector and likewise. Usually the columnized data.}

\item{a}{Numeric. Minimum of the column specified.}

\item{b}{Numeric. Maximum of the column specified.}
}
\value{
An Vector of normalized values.
}
\description{
Normalize an vector
}
