% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_describ.R
\name{print_output}
\alias{print_output}
\title{Print regression summaries into a text file
This function takes one or more fitted model objects.}
\usage{
print_output(..., filepath, text = NULL)
}
\arguments{
\item{...}{One or more fitted model objects
(e.g., objects returned by \code{lm()}, \code{glm()}, or other regression functions).
These are collected into a list and processed sequentially.}

\item{filepath}{Character. The path where the summaries will be written.}

\item{text}{Character default \code{NULL} Text you want to put between models.}
}
\value{
Invisibly returns \code{NULL}. The function is called for its side effect of writing output to an file.
}
\description{
Print regression summaries into a text file
This function takes one or more fitted model objects.
}
