% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_postprocess.R
\name{stdize}
\alias{stdize}
\title{Standardize an vector}
\usage{
stdize(x, mean = NULL, median = FALSE)
}
\arguments{
\item{x}{Vector and likewise. Usually the columnized data.}

\item{mean}{Numeric. Mean of x, na removed.}

\item{median}{Boolean. Whether use median instead of mean.}
}
\value{
An Vector of standardized values.
}
\description{
Standardize an vector
}
