% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_R6.R
\name{TRF_Model}
\alias{TRF_Model}
\title{Class: TRF Model}
\description{
Inherits from \link{BaseBootstrap} to implement TRF-specific bootstrapping.
}
\section{Super class}{
\code{\link[CTRF:BaseBootstrap]{CTRF::BaseBootstrap}} -> \code{TRF_Model}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sim_matrix}}{The generated simulation matrix.}

\item{\code{order}}{Numeric vector. The order of the polynomial and trigonometric terms.}

\item{\code{temp_vars}}{Character vector. Automatically detected temperature-related variables.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TRF_Model-create_sim_matrix}{\code{TRF_Model$create_sim_matrix()}}
\item \href{#method-TRF_Model-run_bootstrap}{\code{TRF_Model$run_bootstrap()}}
\item \href{#method-TRF_Model-aggregate_results}{\code{TRF_Model$aggregate_results()}}
\item \href{#method-TRF_Model-plot}{\code{TRF_Model$plot()}}
\item \href{#method-TRF_Model-clone}{\code{TRF_Model$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CTRF" data-topic="BaseBootstrap" data-id="initialize"><a href='../../CTRF/html/BaseBootstrap.html#method-BaseBootstrap-initialize'><code>CTRF::BaseBootstrap$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRF_Model-create_sim_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-TRF_Model-create_sim_matrix}{}}}
\subsection{Method \code{create_sim_matrix()}}{
Generates the simulation matrix for predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRF_Model$create_sim_matrix(x_range, t_range, order)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_range}}{Numeric vector c(min, max) for the temperature grid.}

\item{\code{t_range}}{Numeric vector c(min, max) for scaling.}

\item{\code{order}}{Numeric vector c(p, q) for model order.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRF_Model-run_bootstrap"></a>}}
\if{latex}{\out{\hypertarget{method-TRF_Model-run_bootstrap}{}}}
\subsection{Method \code{run_bootstrap()}}{
Runs the bootstrap resampling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRF_Model$run_bootstrap(rounds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rounds}}{Integer. Number of bootstrap iterations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRF_Model-aggregate_results"></a>}}
\if{latex}{\out{\hypertarget{method-TRF_Model-aggregate_results}{}}}
\subsection{Method \code{aggregate_results()}}{
Processes bootstrap results into an aggregate table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRF_Model$aggregate_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRF_Model-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TRF_Model-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the median curve and confidence bands.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRF_Model$plot(base_plot, save_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_plot}}{An ggplot2 object to layer the results onto.}

\item{\code{save_path}}{Optional. File path to save the plot.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRF_Model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TRF_Model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRF_Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
