% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcns_postprocess.R
\name{data_finalize}
\alias{data_finalize}
\title{Finalize the necessary modifications}
\usage{
data_finalize(df, hour)
}
\arguments{
\item{df}{Dataframe. The main dataset to work on.
It must contain the complete list of columns.}

\item{hour}{Numeric. Hour.}
}
\value{
An Dataframe augmented with new columns.
}
\description{
Finalize the necessary modifications
}
